Refactor data management of camera plugin
